Moving code related to tags into its own file
