+++ src/ItemNote.ts
@@ -1,6 +1,5 @@
 import log from "loglevel";
 import {
-  App,
   MetadataCache,
   normalizePath,
   Notice,
@@ -170,14 +169,4 @@ export const createOrOpenItemNote =
         return;
       }
     }
-  };
-
-export type OpenSearchForTagFn = (tag: string) => void;
-
-export const openSearchForTag =
-  (app: App): OpenSearchForTagFn =>
-  (tag: string) => {
-    // @ts-ignore
-    const globalSearch = app.internalPlugins.plugins["global-search"].instance;
-    globalSearch.openGlobalSearch(`tag:${tag}`);
   };
